fix(web): place provider installers beside header update controls
